refactor: extract platform timeline builders
